pulling section landing pages into separate modules
