+++ src/films/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Films = () => <h1>Films</h1>;
+
+export default Films;
