+++ src/people/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const People = () => <h1>People</h1>;
+
+export default People;
