+++ src/planets/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Planets = () => <h1>Planets</h1>;
+
+export default Planets;
