+++ src/ships/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Ships = () => <h1>Ships</h1>;
+
+export default Ships;
