+++ src/species/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Species = () => <h1>Species</h1>;
+
+export default Species;
