+++ src/vehicles/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Vehicles = () => <h1>Vehicles</h1>;
+
+export default Vehicles;
